% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_binary.R
\name{backwardPropagation}
\alias{backwardPropagation}
\title{Backward Propagation}
\usage{
backwardPropagation(X, y, cache, params, list_layer_size, use_weights = TRUE)
}
\arguments{
\item{X}{Input feature matrix, dimension (n_x, m)}

\item{y}{True labels (0 or 1), dimension (1, m)}

\item{cache}{Forward propagation cache from \code{\link{forwardPropagation}}}

\item{params}{Parameter list containing W1, b1, W2, b2}

\item{list_layer_size}{List of layer sizes from \code{\link{getLayerSize}}}

\item{use_weights}{Logical, whether to apply class weights for imbalanced data}
}
\value{
A list containing gradients:
\item{dW1}{Gradient of hidden layer weights}
\item{db1}{Gradient of hidden layer bias}
\item{dW2}{Gradient of output layer weights}
\item{db2}{Gradient of output layer bias}
}
\description{
Computes gradients of the loss function with respect to all parameters
using the chain rule.
}
