% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate_Models.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Comprehensive Evaluation for a Single Model}
\usage{
evaluate_models(true_labels, model_result, positive_class)
}
\arguments{
\item{true_labels}{A factor or numeric vector containing the actual labels of the test set.}

\item{model_result}{A list containing the results of a trained model. It must
include \code{model_name}, \code{predictions} (class labels), and \code{pred_prob} (probabilities).}

\item{positive_class}{A character string representing the positive class level
in the target variable (e.g., "1", "Yes", or "X1").}
}
\value{
A data frame containing the computed evaluation metrics.
}
\description{
This function computes core classification metrics (Accuracy, F1-Score,
Precision, Recall, AUC) for a single model's predictions against true test labels.
It also generates visualizations, including an ROC curve and a Confusion Matrix Heatmap.
}
\examples{
\dontrun{
# Assuming 'dtr_result' is the output from decision_tree()
evaluation_df <- evaluate_models(true_labels = y_test,
                                 model_result = dtr_result,
                                 positive_class = "1")
}
}
\author{
Xinyi Hu
}
