% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_binary.R
\name{forwardPropagation}
\alias{forwardPropagation}
\title{Forward Propagation}
\usage{
forwardPropagation(X, params, list_layer_size)
}
\arguments{
\item{X}{Input feature matrix, dimension (n_x, m) where m = number of samples}

\item{params}{Parameter list containing W1, b1, W2, b2 from \code{\link{initializeParameters}}}

\item{list_layer_size}{List of layer sizes from \code{\link{getLayerSize}}}
}
\value{
A list (cache) containing:
\item{Z1}{Linear output of hidden layer, dimension (n_h, m)}
\item{A1}{Hidden layer activation (sigmoid(Z1)), dimension (n_h, m)}
\item{Z2}{Linear output of output layer, dimension (n_y, m)}
\item{A2}{Output layer prediction probabilities, dimension (n_y, m)}
}
\description{
Computes activations for all layers through forward propagation.
Returns all intermediate values for use in backpropagation.
}
\examples{
X <- matrix(rnorm(50), nrow = 5, ncol = 10)
layer_info <- list(n_x = 5, n_h = 4, n_y = 1)
params <- initializeParameters(X, layer_info)
cache <- forwardPropagation(X, params, layer_info)
}
