% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_binary.R
\name{getLayerSize}
\alias{getLayerSize}
\title{Get Layer Sizes of Neural Network}
\usage{
getLayerSize(X, y, hidden_neurons)
}
\arguments{
\item{X}{Input feature matrix. Each column is a sample, each row is a feature.}

\item{y}{Output label matrix. For binary classification, this has 1 row.}

\item{hidden_neurons}{Number of neurons in the hidden layer.}
}
\value{
A list containing:
\item{n_x}{Number of input neurons (feature dimension)}
\item{n_h}{Number of hidden neurons}
\item{n_y}{Number of output neurons}
}
\description{
Determines the number of neurons in each layer based on input and output dimensions.
}
\examples{
X <- matrix(rnorm(100), nrow = 5, ncol = 20)
y <- matrix(sample(0:1, 20, replace = TRUE), nrow = 1)
layer_sizes <- getLayerSize(X, y, hidden_neurons = 4)
}
