% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_binary.R
\name{initializeParameters}
\alias{initializeParameters}
\title{Initialize Neural Network Parameters}
\usage{
initializeParameters(X, list_layer_size)
}
\arguments{
\item{X}{Input feature matrix (used to determine input dimension)}

\item{list_layer_size}{List of layer sizes from \code{\link{getLayerSize}}}
}
\value{
A list containing the initialized parameters:
\item{W1}{Hidden layer weight matrix, dimension (n_h, n_x)}
\item{b1}{Hidden layer bias vector, dimension (n_h, 1)}
\item{W2}{Output layer weight matrix, dimension (n_y, n_h)}
\item{b2}{Output layer bias vector, dimension (n_y, 1)}
}
\description{
Initializes weight matrices using Xavier/Glorot initialization and bias
vectors with zeros. Xavier variance = 2 / (n_input + n_output).
}
\examples{
X <- matrix(rnorm(100), nrow = 5)
layer_info <- list(n_x = 5, n_h = 4, n_y = 1)
params <- initializeParameters(X, layer_info)
}
