% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_binary.R
\name{makePrediction}
\alias{makePrediction}
\title{Make Predictions}
\usage{
makePrediction(X, params, hidden_neurons)
}
\arguments{
\item{X}{Input feature matrix for test data}

\item{params}{Trained parameters from \code{\link{trainModel}}}

\item{hidden_neurons}{Number of neurons in hidden layer (must match training)}
}
\value{
Vector of prediction probabilities in (0,1)
}
\description{
Generates prediction probabilities for new data using trained model.
}
