% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_binary.R
\name{neural_network}
\alias{neural_network}
\title{Neural Network for Binary Classification}
\description{
A collection of functions to build, train, and evaluate a single-hidden-layer
neural network for binary classification tasks. Features include Xavier
parameter initialization, forward/backward propagation, and weighted loss
function for imbalanced data.
}
\author{
LI Zijin
}
