% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_binary.R
\name{NN_performance}
\alias{NN_performance}
\title{Evaluate Model Performance}
\usage{
NN_performance(
  model_name = "Neural Network",
  test_data,
  train_model,
  hidden_neurons
)
}
\arguments{
\item{model_name}{Character string identifying the model}

\item{test_data}{Test feature matrix}

\item{train_model}{Trained model output from \code{\link{trainModel}}}

\item{hidden_neurons}{Number of hidden neurons used in training}
}
\value{
A list containing:
\item{model_name}{Name of the model}
\item{predictions}{Factor vector of predicted classes (0/1)}
\item{pred_prob}{Vector of prediction probabilities}
\item{model}{The trained model object}
}
\description{
Wrapper function to generate predictions and return model performance metrics.
}
