% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForests.R
\name{predict_rf}
\alias{predict_rf}
\title{Random Forest Prediction}
\usage{
predict_rf(trees, X)
}
\arguments{
\item{trees}{List of decision trees from \code{\link{train_rf_scratch}}}

\item{X}{Test feature data frame. Each row is a sample.}
}
\value{
A list containing:
\item{model_name}{Character string "Random Forest"}
\item{predictions}{Factor vector of predicted classes (0/1)}
\item{pred_prob}{Numeric vector of prediction probabilities for class "1"}
\item{model}{The trained random forest model (list of trees)}
}
\description{
Makes predictions for test data using a trained random forest. Returns
both the majority vote class and the probability of belonging to class 1.
}
