% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForests.R
\name{rf_best_split}
\alias{rf_best_split}
\title{Find Best Split}
\usage{
rf_best_split(X, y, mtry)
}
\arguments{
\item{X}{Feature data frame. Each column is a feature, each row is a sample.}

\item{y}{Vector of class labels (0/1).}

\item{mtry}{Number of features randomly sampled as candidates for splitting.}
}
\value{
A list containing:
\item{feature}{Name of the best splitting feature}
\item{value}{Threshold value for the best split}
}
\description{
Randomly selects a subset of features (mtry) and finds the feature-value
split that minimizes the weighted Gini impurity.
}
