% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForests.R
\name{rf_gini_impurity}
\alias{rf_gini_impurity}
\title{Compute Gini Impurity}
\usage{
rf_gini_impurity(y)
}
\arguments{
\item{y}{Vector of class labels (0/1)}
}
\value{
Gini impurity value. Lower values indicate purer nodes.
}
\description{
Calculates the Gini impurity of a set of labels, which measures the
probability of misclassifying a randomly chosen element.
}
\examples{
rf_gini_impurity(c(1,1,1,0,0))  # Returns 0.48
}
