% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM.R
\name{svm_model}
\alias{svm_model}
\title{Support Vector Machine for Binary Classification}
\description{
A collection of functions to train, tune, and evaluate a Support Vector Machine
(SVM) model for binary classification tasks. Features include cross-validation
for hyperparameter tuning and class weighting for imbalanced data.
}
\author{
ZHANG Yibing
}
