% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForests.R
\name{train_rf}
\alias{train_rf}
\title{Train Random Forest from Scratch}
\usage{
train_rf(X, y, n_trees = 20, max_depth = 5)
}
\arguments{
\item{X}{Feature data frame. Each column is a feature, each row is a sample.}

\item{y}{Vector of class labels (0/1).}

\item{n_trees}{Number of trees in the forest. Default is 20.}

\item{max_depth}{Maximum depth of each tree. Default is 5.}
}
\value{
A list of decision trees (length = n_trees)
}
\description{
Builds an ensemble of decision trees using bootstrap aggregation (bagging).
Each tree is trained on a bootstrap sample of the original data.
}
