% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_binary.R
\name{updateParameters}
\alias{updateParameters}
\title{Update Parameters Using Gradient Descent}
\usage{
updateParameters(grads, params, learning_rate)
}
\arguments{
\item{grads}{Gradient list from \code{\link{backwardPropagation}}}

\item{params}{Current parameter list}

\item{learning_rate}{Step size for gradient descent}
}
\value{
Updated parameter list
}
\description{
Updates network parameters by subtracting the gradient times the learning rate.
}
